TST fix tests and get last of code changed
